+++ mb_backend/settings.py
@@ -187,7 +187,7 @@ LOGS_PATH = os.path.join(BASE_DIR, 'logs')
 FILE_SIZE = 1024 * 1024 * 5  # 5 MB
 BACKUP_COUNT = 2
 
-CELERY_BROKER_URL = 'redis://localhost:6379/0'  # or RabbitMQ if preferred
+CELERY_BROKER_URL = 'redis://localhost:6379/0'
 CELERY_ACCEPT_CONTENT = ['json']
 CELERY_TASK_SERIALIZER = 'json'
 
@@ -260,6 +260,14 @@ LOGGING = {
             "backupCount": BACKUP_COUNT,
             "formatter": "verbose",
         },
+        "analytics_handler": {
+            "level": "DEBUG",
+            "class": "logging.handlers.RotatingFileHandler",
+            "filename": os.path.join(LOGS_PATH, "analytics.log"),
+            "maxBytes": FILE_SIZE,
+            "backupCount": BACKUP_COUNT,
+            "formatter": "verbose",
+        },
     },
     "loggers": {
         "django": {
@@ -288,6 +296,11 @@ LOGGING = {
         },
         "Customers": {
             "handlers": ["customers_handler"],
+            "level": "DEBUG",
+            "propagate": False,
+        },
+        "Analytics": {
+            "handlers": ["analytics_handler"],
             "level": "DEBUG",
             "propagate": False,
         },
